1. Controller layer removed,All end points merged in Main 2. UnUsed libraries removed
